[DAD-942] refactor: video 카테고리 파일 분리
