+++ src/lambda_function.py
@@ -8,6 +8,8 @@ import os
 from place import isKakaoPlace, crawlingKakaoPlace
 from place import isNaverPlace, crawlingNaverPlace
 from other import crawlingOther
+from video import isYoutubeVideo, crawlingYoutubeVideo
+from video import isNaverTvVideo, crawlingNaverTvVideo
 
 def lambda_handler(event, context):
     
@@ -24,16 +26,6 @@ def lambda_handler(event, context):
         },
         'body': json.dumps(crawling(url), ensure_ascii=False)
     }
-
-def isYoutubeVideo(url):
-    url_rex = r"(?:https?:\/\/)?(?:www\.)?(?:m\.)?(?:youtube\.com\/(?:watch\?.*v=|shorts\/)|youtu.be\/)([\w-]{11})"
-    url_match = re.search(url_rex, url)
-    return bool(url_match)
-
-def isNaverTvVideo(url):
-    url_rex = r"https:\/\/tv.naver.com\/v\/(\w+)"
-    url_match = re.search(url_rex, url)
-    return bool(url_match)
 
 def isNaverArticle(url):
     url_rex = r"https:\/\/blog.naver.com\/\w+\/\d+"
@@ -122,6 +114,12 @@ def crawling(url):
     if isNaverPlace(url):
         return crawlingNaverPlace(url)
 
+    if isYoutubeVideo(url):
+        return crawlingYoutubeVideo(url)
+    
+    if isNaverTvVideo(url):
+        return crawlingNaverTvVideo(url)
+
 
     result = {}
 
@@ -141,115 +139,8 @@ def crawling(url):
         elif response.encoding.lower() == 'ks_c_5601-1987':
             soup = BeautifulSoup(response.content.decode('ks_c_5601-1987', 'replace'), 'html.parser')
        
-        if isYoutubeVideo(url):
-            
-            #google_api_key 가져오기
-            google_api_key = os.environ.get('GOOGLE_API_KEY', None)
-            
-            #id찾기
-            id_regex = r"(?:https?:\/\/)?(?:www\.)?(?:m\.)?(?:youtube\.com\/(?:watch\?.*v=|shorts\/)|youtu.be\/)([\w-]{11})"
-            id_match = re.search(id_regex, url)
-            id = id_match.group(1)
-
-            #youtube api 호출
-            api_url = "https://www.googleapis.com/youtube/v3/videos?id=" + id + "&key=" + google_api_key +"&part=snippet,statistics,contentDetails"
-            response = requests.get(api_url)
-            json_obj = json.loads(response.text)       
-            
-            #youtube channel api 호출
-            channel_id = json_obj['items'][0]['snippet']['channelId']
-            channel_api_url = "https://www.googleapis.com/youtube/v3/channels?key="+ google_api_key +"&id="+ channel_id +"&part=snippet"
-            channel_response = requests.get(channel_api_url)
-            channel_obj = json.loads(channel_response.text)
-
-            result = {
-                "type" : "video", 
-                "page_url" : url,
-                "embed_url" : "https://www.youtube.com/embed/" + id,
-                "channel_image_url" : None,
-                "site_name" : "YouTube",
-            }
-            
-            try:
-                #게시일 UnixTime으로 변경
-                published_date = json_obj['items'][0]['snippet']['publishedAt']
-                utc_time = datetime.strptime(published_date, "%Y-%m-%dT%H:%M:%SZ")
-                result["published_date"] = int(utc_time.timestamp())
-            except (TypeError, KeyError):
-                result["published_date"] = None
-            
-            try: result["title"] = json_obj['items'][0]['snippet']['title']
-            except (TypeError, KeyError): result["title"] = None
-
-            try: result["thumbnail_url"] = json_obj['items'][0]['snippet']['thumbnails']['high']['url']
-            except (TypeError, KeyError): result["thumbnail_url"] = None
-
-            try: result["description"] = json_obj['items'][0]['snippet']['description']
-            except (TypeError, KeyError): result["description"] = None
-
-            try: result["channel_name"] = json_obj['items'][0]['snippet']['channelTitle']
-            except (TypeError, KeyError): result["channel_name"] = None
-
-            try: result["watched_cnt"] = json_obj['items'][0]['statistics']['viewCount']
-            except (TypeError, KeyError): result["watched_cnt"] = None
-            
-            try: result["channel_image_url"] = channel_obj['items'][0]['snippet']['thumbnails']['high']['url']
-            except (TypeError, KeyError): result["channel_image_url"] = None
-
-            try:
-                duration_str = json_obj['items'][0]['contentDetails']['duration']
-                result['play_time'] = int(parse_duration(duration_str).total_seconds())
-            except (TypeError, KeyError): result["play_time"] = None
-
-            return result
-
-        elif isNaverTvVideo(url):
-
-            result = {
-                "type" : "video",
-                "page_url" : url,
-                "site_name" : "Naver TV",
-            }
-
-            try: result["title"] = soup.select_one('meta[property="og:title"]')['content']
-            except (TypeError, KeyError): result["title"] = None
-
-            try: result["thumbnail_url"] = soup.select_one('meta[property="og:image"]')['content']
-            except (TypeError, KeyError): result["thumbnail_url"] = None
-
-            try: result["description"] = soup.select_one('meta[property="og:description"]')['content']
-            except (TypeError, KeyError): result["description"] = None
-
-            try: result["channel_name"] = soup.select_one('meta[property="og:article:author"]')['content']
-            except (TypeError, KeyError): result["channel_name"] = None
-
-            try: result["channel_image_url"] = soup.select_one('meta[property="og:article:author:image"]')['content']
-            except (TypeError, KeyError): result["channel_image_url"] = None
-
-            try: result["watched_cnt"] = soup.select_one('meta[property="naver:video:play_count"]')['content']
-            except (TypeError, KeyError): result["watched_cnt"] = None
-
-            try: result["play_time"] = soup.select_one('meta[property="naver:video:play_time"]')['content']
-            except (TypeError, KeyError): result["play_time"] = None
-
-            try:
-                published_date = soup.select_one(".date").text.replace('.', '-', 2)[:10]
-                local_datetime = datetime.strptime(published_date, "%Y-%m-%d")
-                result["published_date"] = int(local_datetime.timestamp())
-            except (TypeError, KeyError): result["published_date"] = None 
-            
-            try:
-                #id찾기
-                id_regex = r"https:\/\/tv.naver.com\/v\/(\w+)"
-                id_match = re.search(id_regex, url)
-                id = id_match.group(1)
-                
-                result['embed_url'] = "https://tv.naver.com/embed/" + id
-            except (TypeError, KeyError): result["embed_url"] = None
-                
-            return result
-        
-        elif isNaverArticle(url): 
+    
+        if isNaverArticle(url): 
             #iframe 안에 존재하는 새로운 url 찾기
             redirect_url = "https://blog.naver.com" + soup.select_one('iframe#mainFrame')['src']
